+++ covalent_dispatcher/_cli/service.py
@@ -152,7 +152,7 @@ def _graceful_start(
     develop: bool = False,
 ) -> int:
     """
-    Gracefully start a Flask app.
+    Gracefully start a Fast API app.
 
     Args:
         server_root: Directory where app.py is located.
