+++ covalent_dispatcher/_service/app.py
@@ -19,10 +19,14 @@
 # Relief from the License may be granted by purchasing a commercial license.
 
 import codecs
+import json
 from concurrent.futures import ThreadPoolExecutor
+from typing import Optional
+from uuid import UUID
 
 import cloudpickle as pickle
-from flask import Blueprint, Response, jsonify, make_response, request
+from fastapi import APIRouter, Request
+from fastapi.responses import JSONResponse
 
 import covalent_dispatcher as dispatcher
 from covalent._data_store.datastore import workflow_db
@@ -35,10 +39,11 @@ from .._db.dispatchdb import DispatchDB
 
 app_log = logger.app_log
 log_stack_info = logger.log_stack_info
-bp = Blueprint("dispatcher", __name__, url_prefix="/api")
+
+router: APIRouter = APIRouter()
 
 
-@bp.before_app_first_request
+@router.on_event("startup")
 def start_pools():
     global workflow_pool
     global tasks_pool
@@ -47,8 +52,8 @@ def start_pools():
     tasks_pool = ThreadPoolExecutor()
 
 
-@bp.route("/submit", methods=["POST"])
-def submit() -> Response:
+@router.post("/submit")
+async def submit(request: Request) -> UUID:
     """
     Function to accept the submit request of
     new dispatch and return the dispatch id
@@ -59,18 +64,17 @@ def submit() -> Response:
 
     Returns:
         dispatch_id: The dispatch id in a json format
-                     returned as a Flask Response object.
+                     returned as a Fast API Response object.
     """
+    data = await request.json()
+    data = json.dumps(data).encode("utf-8")
 
-    json_lattice = request.get_data()
-
-    dispatch_id = dispatcher.run_dispatcher(json_lattice, workflow_pool, tasks_pool)
-
-    return jsonify(dispatch_id)
+    dispatch_id = dispatcher.run_dispatcher(data, workflow_pool, tasks_pool)
+    return dispatch_id
 
 
-@bp.route("/cancel", methods=["POST"])
-def cancel() -> Response:
+@router.post("/cancel")
+async def cancel(request: Request) -> str:
     """
     Function to accept the cancel request of
     a dispatch.
@@ -79,70 +83,63 @@ def cancel() -> Response:
         None
 
     Returns:
-        Flask Response object confirming that the dispatch
+        Fast API Response object confirming that the dispatch
         has been cancelled.
     """
-    dispatch_id = request.get_data().decode("utf-8")
+    data = await request.body()
+    dispatch_id = data.decode("utf-8")
 
     dispatcher.cancel_running_dispatch(dispatch_id)
+    return f"Dispatch {dispatch_id} cancelled."
 
-    return jsonify(f"Dispatch {dispatch_id} cancelled.")
 
-
-@bp.route("/db-path", methods=["GET"])
-def db_path() -> Response:
+@router.get("/db-path")
+def db_path() -> str:
     db_path = DispatchDB()._dbpath
-    return jsonify(db_path)
+    return json.dumps(db_path)
 
 
-@bp.route("/result/<dispatch_id>", methods=["GET"])
-def get_result(dispatch_id) -> Response:
+@router.get("/result/{dispatch_id}")
+def get_result(dispatch_id: str, wait: Optional[bool] = True, status_only: Optional[bool] = True):
     app_log.warning("get result")
-    args = request.args
-    wait = args.get("wait", default=False, type=lambda v: v.lower() == "true")
     app_log.warning("wait is " + str(wait))
-    status_only = args.get("status_only", default=False, type=lambda v: v.lower() == "true")
     while True:
         with workflow_db.session() as session:
             lattice_record = (
                 session.query(Lattice).where(Lattice.dispatch_id == dispatch_id).first()
             )
-            status = lattice_record.status if lattice_record else None
-            try:
-                if not lattice_record:
-                    return (
-                        jsonify(
-                            {"message": f"The requested dispatch ID {dispatch_id} was not found."}
-                        ),
-                        404,
-                    )
-                elif not wait or status in [
-                    str(Result.COMPLETED),
-                    str(Result.FAILED),
-                    str(Result.CANCELLED),
-                    str(Result.POSTPROCESSING_FAILED),
-                    str(Result.PENDING_POSTPROCESSING),
-                ]:
-                    output = {
-                        "id": dispatch_id,
-                        "status": lattice_record.status,
-                    }
-                    if not status_only:
-                        output["result"] = codecs.encode(
-                            pickle.dumps(result_from(lattice_record)), "base64"
-                        ).decode()
-                    return jsonify(output)
-
-            except (FileNotFoundError, EOFError):
-                if wait:
-                    continue
-                response = make_response(
-                    jsonify(
-                        {
-                            "message": "Result not ready to read yet. Please wait for a couple of seconds."
-                        }
-                    ),
-                    503,
+        status = lattice_record.status if lattice_record else None
+        try:
+            if not lattice_record:
+                return JSONResponse(
+                    status_code=404,
+                    content={"message": f"The requested dispatch ID {dispatch_id} was not found."},
                 )
-                response.retry_after = 2
-                return response
+            elif not wait or status in [
+                str(Result.COMPLETED),
+                str(Result.FAILED),
+                str(Result.CANCELLED),
+                str(Result.POSTPROCESSING_FAILED),
+                str(Result.PENDING_POSTPROCESSING),
+            ]:
+                output = {
+                    "id": dispatch_id,
+                    "status": lattice_record.status,
+                }
+                if not status_only:
+                    output["result"] = codecs.encode(
+                        pickle.dumps(result_from(lattice_record)), "base64"
+                    ).decode()
+                return output
+
+        except (FileNotFoundError, EOFError):
+            if wait:
+                continue
+            response = JSONResponse(
+                status_code=503,
+                content={
+                    "message": "Result not ready to read yet. Please wait for a couple of seconds."
+                },
+            )
+            response.retry_after = 2
+            return response
