+++ covalent_ui/api/v1/routes/routes.py
@@ -22,7 +22,7 @@
 
 from fastapi import APIRouter
 
-from covalent_dispatcher._service import app_fastapi
+from covalent_dispatcher._service import app
 from covalent_ui.api.v1.routes.end_points import (
     electron_routes,
     graph_route,
@@ -37,4 +37,4 @@ routes.include_router(summary_routes.routes, prefix="/api/v1/dispatches", tags=[
 routes.include_router(lattice_route.routes, prefix="/api/v1/dispatches", tags=["Dispatches"])
 routes.include_router(graph_route.routes, prefix="/api/v1/dispatches", tags=["Graph"])
 routes.include_router(electron_routes.routes, prefix="/api/v1/dispatches", tags=["Electrons"])
-routes.include_router(app_fastapi.router, prefix="/api", tags=["dispatcher"])
+routes.include_router(app.router, prefix="/api", tags=["dispatcher"])
