+++ requirements.txt
@@ -2,9 +2,6 @@ alembic==1.8.0
 cloudpickle==2.0.0
 dask[distributed]==2022.6.0
 fastapi==0.79.0
-flask==2.0.1
-flask-cors==3.0.10
-flask-socketio==5.1.1
 furl==2.1.3
 networkx==2.5
 psutil==5.9.0
